Revert "EXTRA WIP move getViewXml to xmlViewUtils (this is a BREAKING CHANGE)"
This reverts commit 44c446dcfbf604ce4ef0bd7218d594452ac067ac.
