+++ packages/core/src/editor/Editor.ts
@@ -45,8 +45,7 @@ import PrintPreview from '../view/other/PrintPreview.js';
 import Clipboard from '../util/Clipboard.js';
 import MaxLog from '../gui/MaxLog.js';
 import { isNode } from '../util/domUtils.js';
-import { getXml } from '../util/xmlUtils.js';
-import { getViewXml } from '../util/xmlViewUtils.js';
+import { getViewXml, getXml } from '../util/xmlUtils.js';
 import { load, post, submit } from '../util/requestUtils.js';
 import type PopupMenuHandler from '../view/plugin/PopupMenuHandler.js';
 import RubberBandHandler from '../view/plugin/RubberBandHandler.js';
