+++ packages/core/src/index.ts
@@ -180,10 +180,6 @@ export * as styleUtils from './util/styleUtils.js';
  * @category Utils
  */
 export * as xmlUtils from './util/xmlUtils.js';
-/**
- * @category Utils
- */
-export * as xmlViewUtils from './util/xmlViewUtils.js';
 
 export * from './util/config.js';
 export * from './util/logger.js';
