+++ packages/core/src/util/xmlUtils.ts
@@ -16,11 +16,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { NODE_TYPE } from './Constants.js';
+import { NODE_TYPE, NS_SVG } from './Constants.js';
 import { htmlEntities, trim } from './StringUtils.js';
 import { getTextContent } from './domUtils.js';
 import { isElement } from '../internal/utils.js';
 import type { StyleValue } from '../types.js';
+import Point from '../view/geometry/Point.js';
+import type Cell from '../view/cell/Cell.js';
+import type { AbstractGraph } from '../view/AbstractGraph.js';
+import TemporaryCellStates from '../view/cell/TemporaryCellStates.js';
+import Codec from '../serialization/Codec.js';
 
 /**
  * Returns a new, empty XML document.
@@ -31,6 +36,69 @@ export const createXmlDocument = () => {
 
 export const parseXml = (xmlString: string): Document => {
   return new DOMParser().parseFromString(xmlString, 'text/xml');
+};
+
+export const getViewXml = (
+  graph: AbstractGraph,
+  scale = 1,
+  cells: Cell[] | null = null,
+  x0 = 0,
+  y0 = 0
+) => {
+  if (cells == null) {
+    const model = graph.getDataModel();
+    cells = [<Cell>model.getRoot()];
+  }
+
+  const view = graph.getView();
+  let result = null;
+
+  // Disables events on the view
+  const eventsEnabled = view.isEventsEnabled();
+  view.setEventsEnabled(false);
+
+  // Workaround for label bounds not taken into account for image export.
+  // Creates a temporary draw pane which is used for rendering the text.
+  // Text rendering is required for finding the bounds of the labels.
+  const { drawPane } = view;
+  const { overlayPane } = view;
+
+  if (graph.dialect === 'svg') {
+    view.drawPane = document.createElementNS(NS_SVG, 'g');
+    view.canvas.appendChild(view.drawPane);
+
+    // Redirects cell overlays into a temporary container
+    view.overlayPane = document.createElementNS(NS_SVG, 'g');
+    view.canvas.appendChild(view.overlayPane);
+  } else {
+    view.drawPane = <SVGElement>view.drawPane.cloneNode(false);
+    view.canvas.appendChild(view.drawPane);
+
+    // Redirects cell overlays into a temporary container
+    view.overlayPane = <SVGElement>view.overlayPane.cloneNode(false);
+    view.canvas.appendChild(view.overlayPane);
+  }
+
+  // Resets the translation
+  const translate = view.getTranslate();
+  view.translate = new Point(x0, y0);
+
+  // Creates the temporary cell states in the view
+  const temp = new TemporaryCellStates(graph.getView(), scale, cells);
+
+  try {
+    const enc = new Codec();
+    result = enc.encode(graph.getView());
+  } finally {
+    temp.destroy();
+    view.translate = translate;
+    view.canvas.removeChild(view.drawPane);
+    view.canvas.removeChild(view.overlayPane);
+    view.drawPane = drawPane;
+    view.overlayPane = overlayPane;
+    view.setEventsEnabled(eventsEnabled);
+  }
+  return result;
 };
 
 /**
